pyproject.toml
[project]
name = "zotero-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "mcp[cli]>=1.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "pyzotero>=1.6.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
